love2/requirements.txt
# L.O.V.E. Version 2 Requirements

# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.25.0

# AT Protocol (Bluesky)
atproto>=0.0.55

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Optional: Development
ruff>=0.1.0

# Parent project (L.O.V.E. v1) dependencies are assumed to be installed
# in the main requirements.txt
